Refactor structure to tools, reports, and cli handlers
