+++ pd_utils/reports/coverage_gap_report.py
@@ -13,11 +13,6 @@ from pd_utils.model import ScheduleCoverage as SchCoverage
 from pd_utils.util import DateTool
 from pd_utils.util import IOUtil
 from pd_utils.util import PagerDutyQuery
-from pd_utils.util import RuntimeInit
-
-runtime = RuntimeInit("coverage-gap-report")
-runtime.init_secrets()
-runtime.init_logging()
 
 
 class CoverageGapReport:
@@ -146,26 +141,3 @@ class CoverageGapReport:
         for schedule in schedules:
             entries.extend(list(schedule.entries))
         return entries
-
-
-def main(*, _args: list[str] | None = None) -> int:
-    """CLI entry."""
-    runtime.add_standard_arguments(email=False)
-    runtime.add_argument(
-        flag="--look-ahead",
-        default="14",
-        help_="Number of days to look ahead for gaps, default 14",
-    )
-    args = runtime.parse_args(_args)
-    client = CoverageGapReport(token=args.token, look_ahead_days=int(args.look_ahead))
-    schedule_report, escalation_report = client.run_reports()
-
-    now = DateTool.utcnow_isotime().split("T")[0]
-    IOUtil.write_to_file(f"schedule_gap_report{now}.csv", schedule_report)
-    IOUtil.write_to_file(f"escalation_rule_gap_report{now}.csv", escalation_report)
-
-    return 0
-
-
-if __name__ == "__main__":
-    raise SystemExit(main())
